# Fix: Refs #207 Fix implementation of Jensen Shannon measure