--- a/include/simsimd/probability.h
+++ b/include/simsimd/probability.h
@@ -108,7 +108,7 @@ SIMSIMD_PUBLIC void simsimd_js_f16_sapphire(simsimd_f16_t const* a, simsimd_f16_
             d += ai * SIMSIMD_LOG((ai + epsilon) / (mi + epsilon));                                           \
             d += bi * SIMSIMD_LOG((bi + epsilon) / (mi + epsilon));                                           \
         }                                                                                                     \
-        *result = SIMSIMD_SQRT(((simsimd_distance_t)d / 2));                                                  \  
+        *result = SIMSIMD_SQRT(((simsimd_distance_t)d / 2));                                                  \
     }
 
 SIMSIMD_MAKE_KL(serial, f64, f64, SIMSIMD_DEREFERENCE, SIMSIMD_F32_DIVISION_EPSILON) // simsimd_kl_f64_serial
@@ -198,8 +198,6 @@ SIMSIMD_PUBLIC void simsimd_js_f32_neon(simsimd_f32_t const *a, simsimd_f32_t co
     simsimd_f32_t epsilon = SIMSIMD_F32_DIVISION_EPSILON;
     float32x4_t epsilon_vec = vdupq_n_f32(epsilon);
     float32x4_t sum_vec = vdupq_n_f32(0);
-    float32x4_t half_vec = vdupq_n_f32(0.5f);
-
     float32x4_t a_vec, b_vec;
 
 simsimd_js_f32_neon_cycle:
@@ -226,7 +224,7 @@ simsimd_js_f32_neon_cycle:
     if (n != 0) goto simsimd_js_f32_neon_cycle;
 
     simsimd_f32_t log2_normalizer = 0.693147181f;
-    sum_vec = vmulq_f32(sum_vec, half_vec);
+    sum_vec = vmulq_n_f32(sum_vec, 0.5f);
     simsimd_f32_t sum = vaddvq_f32(sum_vec) * log2_normalizer;
     *result = SIMSIMD_SQRT(sum);
 }
@@ -275,8 +273,6 @@ SIMSIMD_PUBLIC void simsimd_js_f16_neon(simsimd_f16_t const *a, simsimd_f16_t co
     float32x4_t sum_vec = vdupq_n_f32(0);
     simsimd_f32_t epsilon = SIMSIMD_F32_DIVISION_EPSILON;
     float32x4_t epsilon_vec = vdupq_n_f32(epsilon);
-    float32x4_t half_vec = vdupq_n_f32(0.5f);
-
     float32x4_t a_vec, b_vec;
 
 simsimd_js_f16_neon_cycle:
@@ -296,13 +292,13 @@ simsimd_js_f16_neon_cycle:
     float32x4_t ratio_b_vec = vdivq_f32(vaddq_f32(b_vec, epsilon_vec), vaddq_f32(m_vec, epsilon_vec));
     float32x4_t log_ratio_a_vec = _simsimd_log2_f32_neon(ratio_a_vec);
     float32x4_t log_ratio_b_vec = _simsimd_log2_f32_neon(ratio_b_vec);
-    float32x4_t prod_a_vec = vmulq_f32(vmulq_f32(a_vec, log_ratio_a_vec), half_vec);
-    float32x4_t prod_b_vec = vmulq_f32(vmulq_f32(b_vec, log_ratio_b_vec), half_vec);
+    float32x4_t prod_a_vec = vmulq_f32(a_vec, log_ratio_a_vec);
+    float32x4_t prod_b_vec = vmulq_f32(b_vec, log_ratio_b_vec);
     sum_vec = vaddq_f32(sum_vec, vaddq_f32(prod_a_vec, prod_b_vec));
     if (n) goto simsimd_js_f16_neon_cycle;
 
     simsimd_f32_t log2_normalizer = 0.693147181f;
-    sum_vec = vmulq_f32(sum_vec, half_vec);
+    sum_vec = vmulq_n_f32(sum_vec, 0.5f);
     simsimd_f32_t sum = vaddvq_f32(sum_vec) * log2_normalizer;
     *result = SIMSIMD_SQRT(sum);
 }