The Jensen Shannom measure implementation in SimSIMD is different from the implementation in scipy. It turns out that a scaling factory of 0.5 was missed and this fix seems to match the example provided in scipy documentation.

Matches the examples provided in the [scipy documentation](https://github.com/scipy/scipy/blob/v1.14.1/scipy/spatial/distance.py#L1186-L1271)