@ashvardanian 2024-11-17
Here is the numerical error report:

```
+---+---------------+------+---------+---------------------+---------------------+---------------------+---------------------+---------------------+-----------------+
|   |    Metric     | NDim |  DType  |   Baseline Error    |    SimSIMD Error    |  Accurate Duration  |  Baseline Duration  |  SimSIMD Duration   | SimSIMD Speedup |
+---+---------------+------+---------+---------------------+---------------------+---------------------+---------------------+---------------------+-----------------+
| 0 | jensenshannon |  11  | float16 | 3.03e-04 ± 2.79e-04 | 8.61e-02 ± 1.53e-02 | 1.91e+04 ± 1.58e+03 | 1.93e+04 ± 1.36e+03 | 1.16e+03 ± 4.91e+02 | 17.32x ± 2.11x  |
| 1 | jensenshannon |  11  | float32 | 4.47e-08 ± 4.07e-08 | 7.44e-02 ± 3.15e-02 | 1.95e+04 ± 2.48e+03 | 1.76e+04 ± 3.68e+03 | 1.18e+03 ± 8.24e+02 | 16.39x ± 3.24x  |
| 2 | jensenshannon |  97  | float16 | 1.08e-04 ± 8.37e-05 | 9.11e-02 ± 4.34e-03 | 2.14e+04 ± 1.82e+03 | 2.33e+04 ± 1.58e+03 | 1.41e+03 ± 3.29e+02 | 16.72x ± 1.60x  |
| 3 | jensenshannon |  97  | float32 | 3.07e-08 ± 2.18e-08 | 7.99e-02 ± 3.03e-02 | 2.15e+04 ± 1.93e+03 | 1.88e+04 ± 1.17e+03 | 1.20e+03 ± 5.35e+02 | 16.31x ± 1.89x  |
| 4 | jensenshannon | 1536 | float16 | 8.90e-05 ± 6.03e-05 | 9.22e-02 ± 9.60e-04 | 5.09e+04 ± 2.73e+03 | 9.97e+04 ± 3.40e+03 | 6.08e+03 ± 5.49e+02 | 16.47x ± 0.89x  |
| 5 | jensenshannon | 1536 | float32 | 2.40e-08 ± 1.81e-08 | 8.07e-02 ± 3.03e-02 | 5.07e+04 ± 3.08e+03 | 4.74e+04 ± 1.22e+04 | 3.56e+03 ± 7.67e+02 | 13.51x ± 2.89x  |
+---+---------------+------+---------+---------------------+---------------------+---------------------+---------------------+---------------------+-----------------+
```