@ashvardanian 2024-11-17
Hi, @GoWind! It seems like you didn't take care of the x86 code. It's also not recommended to use the `SIMSIMD_SQRT` in the vectorized code. There are native approximations for those.