@GoWind 2024-11-17
Hi @ashvardanian , I am working on a follow up PR for the intel versions, will submit them in a day or so.

Regd, `the SIMSIMD_SQRT` marcro, i see that there is a `*result = _simsimd_sqrt_f64_neon(sum)` that returns an f64_t (which `simsimd_distance_t` is a typedef for). Can I use them  in the js_f32_neon and js_16_neon implementations because *result is of type simsimd_distance_t anyway ?