@ashvardanian 2024-11-17
> Can I use them in the js_f32_neon and js_16_neon implementations because *result is of type simsimd_distance_t anyway ?

Yes, @GoWind. I've already patched that part and will probably merge x86 patches soon as well. Would you like to look into #187 or the simpler #202, #203, or #228?