@GoWind 2024-11-17
@ashvardanian , glad to :) #187  is marked as closed, does it still need a fix ? 

I will try to fix 202 or 203 as well and then see how I can contribute to the kernel fn documentation for C :)